move src/copy.rs -> src/io/copy.rs
