+++ tokio-io/src/lib.rs
@@ -49,7 +49,6 @@ pub mod io;
 pub mod codec;
 
 mod allow_std;
-mod copy;
 mod framed;
 mod framed_read;
 mod framed_write;
